+++ graphite/ecs/MeshComponent.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <string>
-
-struct MeshComponent
-{
-    std::string primitiveID;
-};