+++ graphite/ecs/RenderableComponent.h
@@ -1,0 +1,14 @@
+#pragma once
+
+// #include <string>
+#include <filesystem>
+
+using AssetID = std::filesystem::path;
+
+struct RenderableComponent
+{
+    // std::string primitiveID;
+
+    AssetID meshID;
+    AssetID materialID;
+};