+++ graphite/rendering/Material.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <filesystem>
+
+using AssetID = std::filesystem::path;
+
+struct Material
+{
+    AssetID albedo;
+    AssetID normal;
+    AssetID orm;
+
+    float baseColor[4] = {1, 1, 1, 1};
+    float roughness = 1.0f;
+    float metallic = 0.0f;
+};